H10 re-bank i9 W=4 slice 2: 75 of 76 rows at c6fdbe73c3
Test artifacts only, per the 2026-09-22 bank ruling: *_test.cs, package_test_info.cs,
package_info_internal_test.cs, package_init_internal_test.cs, go2cs_test_host.cs, <pkg>.tests.csproj,
the proof pages of the PASS rows, the modified README badges, and roster edits for 11 PASS rows.
No production .cs, package_info.cs or production .csproj. Production files restored in the leg
tree: 40. The time row (BUILD, banked=no) contributes nothing. Not banked: GOROOT testdata mirrored
into sibling packages, docs/validation/index.md; roster header not touched.

Words: 74 PASS, 1 DIVERGED (hash/maphash), 1 BUILD (time, not banked).

+++ src/core/bufio/scan_test.cs
@@ -194,7 +194,7 @@ public static void TestScanLongLines(ж<Δtesting.T> Ꮡt) {
     var buf = @new<bytes.Buffer>();
     nint lineNum = 0;
     nint j = 0;
-    for (nint i = 0; i < 2 * smallMaxTokenSize; i++) {
+    for (nint i = 0; i < (nint)(2 * smallMaxTokenSize); i++) {
         genLine(tmp, lineNum, j, true);
         if (j < smallMaxTokenSize){
             j++;
@@ -234,7 +234,7 @@ public static void TestScanLineTooLong(ж<Δtesting.T> Ꮡt) {
     var buf = @new<bytes.Buffer>();
     nint lineNum = 0;
     nint j = 0;
-    for (nint i = 0; i < 2 * smallMaxTokenSize; i++) {
+    for (nint i = 0; i < (nint)(2 * smallMaxTokenSize); i++) {
         genLine(tmp, lineNum, j, true);
         j++;
         buf.Write(tmp.Bytes());
